Source/Draft: A prototype implementation of mini git downloader
    Setup a thread and a process. Setup communication pipes
amongst them. The thread is responsible for fetching packfile and
writing to tmp_file (fetch_XXXXXX). It then forwards the tmp_file
content to the main thread. The main thread processes the
packfile which is encoded in pkt-line format, and forwards it to
its child process. This process is responsible with deserializing
it into git object types(only applicable onces are processed):
- OBJ_TREE
- OBJ_COMMIT
- OBJ_BLOB

Limitations:

 * We specifically request the server for OBJ_REF_DELTA format.
   It is less complicated to handle compared to OBJ_OFS_DELTA.

 * Only projects using SHA1 hash are supported.

 * This draft implementation crashes if the required SHA1 for
   the object type it requires to build its directory is encoded
   in OBJ_REF_DELTA.

Improvement:

    While we can't escape the creation of the tmp_file since we
have to create an index file which is required to resolve the
OBJ_*_DELTA types, we can use a linked list structure to hold the
packets as we receive them. This can then be queued appropriately
to the channel if it is not busy. This way we don't have to wait
to download all packfile, we can parallelize the interpretation
of the packfile routine.