Source/Minor: Implement fan-out table
Use a fan-out index table to reduce the
number of hash entry comparison required
to find one. This is a 4 bytes 256 entries
lookup table with the elements containing
prefix sum of the number of occurrence of
all elements less than or equal to the first
hash index of the hash entry.
Illustration:

hash value: 4b1c2c30f...
Key: 4b(75)
fan_out[75] contains the sum of all occurrence
of `4b` and less.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>