Source/feat: Full support for delta resolution and streaming
All required deltas for the generation of a
working tree are now fully resolved.

The fan-out table idea has been abandoned as
we would need to insert a newly resolved hash
into the objects structure. This requires us
to rebuild/expand the object structure, sort
it on every delta resolved. This is impractical.

The replacement designed is to use an hash map.
We hash all non-delta objects into an obj_map.
For every delta resolved, we add the hash(oid)
of the just resolved delta as a new entry in
the map. This way, any new delta that depends
on the just resolved delta can use it as a base.

Hurray!, we now process the objects in parallel
as we are receiving it from the fetching thread.
A new structure vtask has been added to achieve
this. The main thread consumes the received
packet via the *from_head functions. We keep
track of the total number of packets we have seen
so far, and consume it only when we have at least
the exact number of bytes requested by the
decoded pkt-line message. This way, we know that
we don't get truncated packet.

On the packet receiving thread, we adapt
check_pktline routine from git source. This helps
in validating the pkt-line we just received.

An extra argument has been added (deepen 1) to
consume only the required packet for the latest
revision. This way we make sure we are not cloning
the project, we are downloading and not waste user's
data.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>