Source/Minor: Improve error checking
To notify the parent thread when the fetching thread
dies we add a new variable task_fail. The parent
thread extends its waiting thread with this variable.
It checks to see if it is not null. If it is, then
the fetching thread has failed unexpectedly. We
should exit also.

We add the check for to compute SHA of all data
used by the make_tree process and we validate
against the one send at the close to the end of
transmission by the server.

We use CURLOPT_*TIMEOUT variables to set the maximum
time for waiting thread to stay alive when the
server's response speed drops below 30bytes for up to
1 minute.

We improve dependency management by the use of CMake
Package Manager(CPM). With this, we are assured to
give a better developer experience.

Signed-off-by: Meekness Adesina <zenon8adams@gmail.com>